Changed deprecation mechanism
The factory checks with the VersionedNodeSetFactoryManager whether it
is deprecated or not. This avoids writing the info into the additional
settings, which is wrong because that info is persisted. I also removed
the VersionedNodeSetFactory because that stuff can be handled by the
DynamicGenericNodeSetFactory, which now by default gets its version
from the plugin. And finally I fixed the handling of paths  with spaces.